Better handling of situation where logfile can't be opened
If freopen() fails, stderr is diverted to an undisclosed location
and it's not clear what is going on.

Also add an explicit notice announcing our intention to divert
logging output to a file.

Per #105.

Note that it might make sense to disable logfile output when
running the repmgr command line client as normally you'd expect
immediate feedback.